\name{StatComp22084-package}
\alias{StatComp22084-package}
\alias{StatComp22084}
\docType{package}
\title{
  This package provides several functions for computing LDPE and RLDPE.
}
\description{
  This package provides several functions for computing LDPE and RLDPE.
}

\author{
22084

Maintainer: 22084 <shizhiheng@mail.ustc.edu.cn>
}

\keyword{ package }



