\name{StatComp22084-package}
\alias{StatComp22084-package}
\alias{StatComp22084}
\docType{package}
\title{
   Homework and research related functions inside
}
\description{
   Homework and research related functions inside
}
\details{
   Homework and research related functions insides.
}
\author{
22084

Maintainer: 22084 <shizhiheng@mail.ustc.edu.cn>
}

\keyword{ package }


